+++ attendees/whereabouts/models/campus.py
@@ -71,7 +71,7 @@ class CampusesHistory(pghistory.get_event_model(
     pgh_created_at = models.DateTimeField(auto_now_add=True)
     pgh_label = models.TextField(help_text='The event label.')
     pgh_obj = models.ForeignKey(db_constraint=False, on_delete=models.deletion.DO_NOTHING, related_name='history', to='whereabouts.campus')
-    id = models.IntegerField()
+    id = models.BigIntegerField()
     created = model_utils.fields.AutoCreatedField(default=django.utils.timezone.now, editable=False, verbose_name='created')
     modified = model_utils.fields.AutoLastModifiedField(default=django.utils.timezone.now, editable=False, verbose_name='modified')
     is_removed = models.BooleanField(default=False)
