+++ attendees/whereabouts/models/room.py
@@ -64,7 +64,7 @@ class RoomsHistory(pghistory.get_event_model(
     modified = model_utils.fields.AutoLastModifiedField(default=django.utils.timezone.now, editable=False, verbose_name='modified')
     is_removed = models.BooleanField(default=False)
     slug = models.SlugField(db_index=False)
-    id = models.IntegerField()
+    id = models.BigIntegerField()
     infos = models.JSONField(blank=True, default=dict, help_text='Example: {"accessibility": 3}. Please keep {} here even no data', null=True)
     suite = models.ForeignKey(db_constraint=False, null=True, on_delete=models.deletion.DO_NOTHING, related_name='+', related_query_name='+', to='whereabouts.suite')
     display_name = models.CharField(max_length=50)
